+++ samples/android-app/src/main/kotlin/org/jlleitschuh/gradle/ktlint/android/AlternativeSample.kt
@@ -1,0 +1,5 @@
+package org.jlleitschuh.gradle.ktlint.android
+
+data class AlternativeSample(
+    val one: Int = 1
+)